Many changes

+++ __main__.py
@@ -1,12 +1,13 @@
 import asyncio
 import multiprocessing
 import subprocess
+from typing import List
 import uvicorn
 
-from sqlalchemy import text
 from r_mutex import Lock
-from engine.futures_engine import FuturesEngine
-from engine.pusher import Pusher
+from sqlalchemy import select, text
+
+from api.routes.instrument.utils import cache_market_data
 from config import (
     INSTRUMENT_LOCK_PREFIX,
     ORDER_LOCK_PREFIX,
@@ -15,8 +16,18 @@ from config import (
     DEV_MODE,
     DB_URL,
 )
+from db_models import Instruments
+from engine.futures_engine import FuturesEngine
+from engine.pusher import Pusher
 from engine.spot_engine import SpotEngine
 from utils.db import get_db_session, remove_sqlalchemy_url, write_sqlalchemy_url
+
+
+async def fetch_instruments() -> list[str]:
+    async with get_db_session() as sess:
+        res = await sess.execute(select(Instruments.instrument))
+        res = res.all()
+    return [item[0] for item in res]
 
 
 async def handle_run_server() -> None:
@@ -26,7 +37,7 @@ async def handle_run_server() -> None:
     fa_config = uvicorn.Config(
         "api.app:app",
         workers=3,
-        host="0.0.0.0", 
+        host="0.0.0.0",
         port=8000,
         # log_config=None,
     )
@@ -41,27 +52,51 @@ def run_server() -> None:
     asyncio.run(handle_run_server())
 
 
-async def handle_run_engine() -> None:
-    """Handles configs for engine and running it"""
+async def handle_run_engine(engine_class) -> None:
     order_lock = Lock(REDIS_CLIENT, ORDER_LOCK_PREFIX, is_manager=False)
     instrument_lock = Lock(REDIS_CLIENT, INSTRUMENT_LOCK_PREFIX)
 
     pusher = Pusher(order_lock)
-    futures_engine = FuturesEngine(instrument_lock, pusher)
-    spot_engine = SpotEngine(instrument_lock, pusher)
-
     asyncio.create_task(order_lock.run())
     asyncio.create_task(instrument_lock.run())
 
-    await asyncio.gather(*[futures_engine.run(), spot_engine.run()])
+    engine = engine_class(instrument_lock, pusher)
+    await engine.run(await fetch_instruments())
 
 
-def run_engine() -> None:
-    """Engine Entrypoint"""
-    asyncio.run(handle_run_engine())
+def run_engine(engine_class):
+    asyncio.run(handle_run_engine(engine_class))
 
 
-async def migrate() -> None:
+async def listen_for_instruments(instruments: List[str], lock: asyncio.Lock) -> None:
+    async with REDIS_CLIENT.pubsub() as ps:
+        await ps.subscribe("instrument.new")
+        async for message in ps.listen():
+            if message["type"] == "subscribe":
+                continue
+
+            async with lock:
+                instruments.append(message["data"])
+
+
+async def handle_market_data_cache():
+    lock = asyncio.Lock()
+    instruments = await fetch_instruments()
+    sleep = 60 * 60 * 2  # seconds
+    asyncio.create_task(listen_for_instruments(instruments, lock))
+
+    while True:
+        async with lock:
+            for instrument in instruments:
+                await cache_market_data(instrument)
+        await asyncio.sleep(sleep)
+
+
+def run_market_data_cache():
+    asyncio.run(handle_market_data_cache())
+
+
+async def run_migrate() -> None:
     """Runs alembic commands. To be run before before running server"""
     write_sqlalchemy_url(DB_URL.replace("+asyncpg", ""))
     subprocess.run(["alembic", "upgrade", "head"], check=True)
@@ -69,7 +104,7 @@ async def migrate() -> None:
     async with get_db_session() as sess:
         if DEV_MODE:
             await sess.execute(text("DELETE FROM orders;"))
-            await sess.execute(text("DELETE FROM users;"))
+            await sess.execute(text("DELETE FROM users WHERE username != 'admin';"))
 
         await sess.execute(text("SELECT COUNT(*) FROM orders;"))
         await sess.commit()
@@ -83,13 +118,15 @@ async def main() -> None:
 
     ps = [
         multiprocessing.Process(target=run_server, name="server"),
-        multiprocessing.Process(target=run_engine, name="engine"),
+        multiprocessing.Process(target=run_engine, args=(FuturesEngine,), name="futures engine"),
+        multiprocessing.Process(target=run_engine, args=(SpotEngine,), name="spot engine"),
+        multiprocessing.Process(target=run_market_data_cache, name="market data cache"),
     ]
 
     for p in ps:
         p.start()
 
-    await migrate()
+    await run_migrate()
 
     try:
         while True:
